+++ src/Actions/SendMailAction.php
@@ -4,19 +4,24 @@ namespace Binarcode\LaravelMailator\Actions;
 
 use Binarcode\LaravelMailator\Events\ScheduleMailSentEvent;
 use Binarcode\LaravelMailator\Models\MailatorSchedule;
+use Binarcode\LaravelMailator\Support\ClassResolver;
 use Exception;
 use Illuminate\Support\Facades\Mail;
 
 class SendMailAction implements Action
 {
+    use ClassResolver;
+
     public function handle(MailatorSchedule $schedule)
     {
         try {
             $this->sendMail($schedule);
-        } catch (Exception $exception) {
-            report($exception);
 
+            static::garbageResolver()->handle($schedule);
+        } catch (Exception $exception) {
             $schedule->markAsFailed($exception->getMessage());
+
+            report($exception);
         }
     }