+++ src/LaravelMailatorServiceProvider.php
@@ -2,6 +2,8 @@
 
 namespace Binarcode\LaravelMailator;
 
+use Binarcode\LaravelMailator\Console\Commands\GarbageCollectorCommand;
+use Binarcode\LaravelMailator\Console\Commands\MailatorSchedulerCommand;
 use Binarcode\LaravelMailator\Models\MailatorLog;
 use Binarcode\LaravelMailator\Models\MailatorSchedule;
 use Binarcode\LaravelMailator\Models\MailTemplate;
@@ -65,7 +67,10 @@ class LaravelMailatorServiceProvider extends ServiceProvider
             ], 'lang');*/
 
             // Registering package commands.
-            // $this->commands([]);
+            $this->commands([
+                 GarbageCollectorCommand::class,
+                 MailatorSchedulerCommand::class,
+             ]);
         }
     }
 
@@ -79,7 +84,11 @@ class LaravelMailatorServiceProvider extends ServiceProvider
 
         // Register the main class to use with the facade
         $this->app->singleton('mailator', function () {
-            return new LaravelMailator();
+            return new MailatorManager();
+        });
+
+        $this->app->singleton('mailator-scheduler', function () {
+            return new SchedulerManager();
         });
     }
 }