+++ src/Models/MailatorLog.php
@@ -7,6 +7,7 @@ use Illuminate\Database\Eloquent\Model;
 
 /**
  * Class MailatorLog
+ * @property string $status
  * @property Carbon $created_at
  * @package Binarcode\LaravelMailator\Models
  */
@@ -17,8 +18,8 @@ class MailatorLog extends Model
         return config('mailator.logs_table', 'mailator_logs');
     }
 
-    const STATUS_FAILED = 'failed';
-    const STATUS_SENT = 'sent';
+    public const STATUS_FAILED = 'failed';
+    public const STATUS_SENT = 'sent';
 
     protected $fillable = [
         'name',
@@ -41,5 +42,10 @@ class MailatorLog extends Model
     public function isSent(): bool
     {
         return $this->status === static::STATUS_SENT;
+    }
+
+    public function isFailed(): bool
+    {
+        return $this->status === static::STATUS_FAILED;
     }
 }