+++ src/Support/ConverterEnum.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Binarcode\LaravelMailator\Support;
+
+class ConverterEnum
+{
+    public const MINUTES_IN_HOUR = 60;
+    public const MINUTES_IN_DAY = 60 * 60;
+    public const MINUTES_IN_WEEK = 168 * 60;
+    public const HOURS_IN_DAY = 24;
+    public const HOURS_IN_WEEK = 168;
+}