+++ tests/Fixtures/Constraints/TrueConstraint.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Binarcode\LaravelMailator\Tests\Fixtures;
+namespace Binarcode\LaravelMailator\Tests\Fixtures\Constraints;
 
 use Binarcode\LaravelMailator\Constraints\SendScheduleConstraint;
 use Binarcode\LaravelMailator\Models\MailatorSchedule;
 use Illuminate\Support\Collection;
 
-class BeforeInvoiceExpiresConstraint implements SendScheduleConstraint
+class TrueConstraint implements SendScheduleConstraint
 {
-    public function canSend(MailatorSchedule $schedule, Collection $log): bool
+    public function canSend(MailatorSchedule $schedule, Collection $logs): bool
     {
         return true;
     }